backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
anthropic
httpx
faker
tenacity
tqdm
